ArchMCP-Drawio/requirements.txt
streamlit>=1.28.0
diagrams>=0.23.4
graphviz>=0.20.1
pillow>=10.0.0
boto3>=1.34.0
botocore>=1.34.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
mermaid-py>=0.3.0
